pyproject.toml
[build-system]
requires = ["setuptools>=63", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pseudo-tv"
version = "0.1.0"
description = "Standalone pseudo TV simulator that builds virtual channels from a media library"
readme = "readme.md"
authors = [{ name = "Pseudo TV Contributors" }]
requires-python = ">=3.9"
license = { file = "LICENSE" }
dependencies = []
keywords = ["kodi", "pseudo tv", "media", "scheduler", "epg"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
yaml = ["PyYAML>=6.0"]

[project.urls]
Homepage = "https://example.com/pseudo-tv"
Source = "https://example.com/pseudo-tv/source"

[tool.setuptools.packages.find]
include = ["pseudo_tv"]
